# Tidy up HttpFake