--- a/src/libaktualizr/primary/aktualizr_test.cc
+++ b/src/libaktualizr/primary/aktualizr_test.cc
@@ -14,12 +14,14 @@
 #include "libaktualizr/events.h"
 
 #include "httpfake.h"
+#include "metafake.h"
 #include "primary/aktualizr_helpers.h"
 #include "primary/sotauptaneclient.h"
 #include "uptane_test_common.h"
 #include "utilities/utils.h"
 #include "virtualsecondary.h"
 
+#include "uptane_repo.h"
 #include "utilities/fault_injection.h"
 
 boost::filesystem::path uptane_repos_dir;
@@ -2365,7 +2367,7 @@ int main(int argc, char** argv) {
 
   TemporaryDirectory tmp_dir;
   fake_meta_dir = tmp_dir.Path();
-  MetaFake meta_fake(fake_meta_dir);
+  CreateFakeRepoMetaData(fake_meta_dir);
 
   return RUN_ALL_TESTS();
 }