--- a/src/libaktualizr/primary/metadata_fetch_test.cc
+++ b/src/libaktualizr/primary/metadata_fetch_test.cc
@@ -85,9 +85,9 @@ TEST(Aktualizr, MetadataFetch) {
 
   // Two images added, but only one update scheduled: all metadata objects
   // should be fetched once.
-  uptane_repo_.addImage("tests/test_data/firmware.txt", "firmware.txt", "primary_hw");
+  uptane_repo_.addImage("tests/test_data/firmware.txt", "firmware-małecki.txt", "primary_hw");
   uptane_repo_.addImage("tests/test_data/firmware_name.txt", "firmware_name.txt", "primary_hw");
-  uptane_repo_.addTarget("firmware.txt", "primary_hw", "CA:FE:A6:D2:84:9D");
+  uptane_repo_.addTarget("firmware-małecki.txt", "primary_hw", "CA:FE:A6:D2:84:9D");
   uptane_repo_.addDelegation(Uptane::Role("role-abc", true), Uptane::Role("targets", false), "abc/*", false,
                              KeyType::kED25519);
   uptane_repo_.signTargets();
@@ -123,7 +123,7 @@ TEST(Aktualizr, MetadataFetch) {
   // Delegation added to an existing delegation; update scheduled with
   // pre-existing image: Snapshot must be refetched, but Targets are unchanged.
   uptane_repo_.emptyTargets();
-  uptane_repo_.addTarget("firmware.txt", "primary_hw", "CA:FE:A6:D2:84:9D");
+  uptane_repo_.addTarget("firmware-małecki.txt", "primary_hw", "CA:FE:A6:D2:84:9D");
   uptane_repo_.addDelegation(Uptane::Role("role-def", true), Uptane::Role("role-abc", true), "def/*", false,
                              KeyType::kED25519);
   uptane_repo_.signTargets();