--- a/src/libaktualizr/uptane/uptane_test.cc
+++ b/src/libaktualizr/uptane/uptane_test.cc
@@ -13,6 +13,7 @@
 #include <boost/filesystem.hpp>
 #include "json/json.h"
 
+#include "crypto/crypto.h"
 #include "crypto/p11engine.h"
 #include "httpfake.h"
 #include "libaktualizr/secondaryinterface.h"