I've been looking into a problem with targets that contain non-ascii characters.  While doing the investigation work, I tidied up a few places, which formed this PR.

For background, Aktualizr and libtuf have different ideas about how to canonicalise JSON. The details are in my [multibyte](https://github.com/cajun-rat/aktualizr/tree/multibyte) branch, with [this commit](https://github.com/uptane/aktualizr/commit/271e6a04394c1e6319a0663b0de0fbed2687c54dl) being the critical one that would change things.

For now, I hope these changes are benign and will be useful in general.