@pattivacek 2023-02-14
> > Were the header changes done just to speed up the tests?
> 
> Mostly it is because I'm allergic to implementations in header files, and the include file bloat that it brings. I did some quick tests by setting `CCACHE:BOOL=OFF` in CMakeCache and running `ninja clean && perf stat ninja -j 4 build_tests`, and the results were:

Nice, I love numbers! I mean, even if there were no speedup, I have no objection at all to the changes. I'm quite happy to get things out of headers.